move the server part to utils.ts in login paage
