+++ src/app/login/utils.ts
@@ -1,0 +1,16 @@
+import { createClient } from '@/lib/supabase/client';
+
+const supabase = createClient();
+
+export const handleGoogleLogin = async () => {
+    await supabase.auth.signInWithOAuth({
+      provider: 'google',
+      options: {
+        redirectTo: `${window.location.origin}/auth/callback`,
+        queryParams: {
+          access_type: 'offline',
+          prompt: 'consent',
+        },
+      },
+    });
+};
